int to float in a function

+++ src/BoardLogic.cpp
@@ -74,8 +74,8 @@ bool BoardLogic::TryAddNewColumn()
 	tiles.emplace_back();
 	for (int rowIndex = 0; rowIndex < BOARD_COLUMN_BLOCK_NUMBER; ++rowIndex)
 	{
-		float xPos = WINDOW_WIDTH - TILE_SIDE;
-		float yPos = WINDOW_HEIGHT - BOARD_INITIALCOLUMN_HEIGHT_POS - TILE_SIDE - rowIndex * TILE_SIDE;
+		int xPos = WINDOW_WIDTH - TILE_SIDE;
+		int yPos = WINDOW_HEIGHT - BOARD_INITIALCOLUMN_HEIGHT_POS - TILE_SIDE - rowIndex * TILE_SIDE;
 		tiles[tiles.size()-1].emplace_back(xPos, yPos, TILE_SIDE, GetRandomColor());
 	}
 	return true;
@@ -90,8 +90,8 @@ void BoardLogic::ResetBoard()
 		tiles.emplace_back();
 		for (int rowIndex = 0; rowIndex < BOARD_COLUMN_BLOCK_NUMBER; ++rowIndex)
 		{
-			float xPos = WINDOW_WIDTH - ((BOARD_GAMESTART_COLUMNS - columnIndex) * TILE_SIDE);
-			float yPos = WINDOW_HEIGHT - BOARD_INITIALCOLUMN_HEIGHT_POS - TILE_SIDE - rowIndex * TILE_SIDE;
+			int xPos = WINDOW_WIDTH - ((BOARD_GAMESTART_COLUMNS - columnIndex) * TILE_SIDE);
+			int yPos = WINDOW_HEIGHT - BOARD_INITIALCOLUMN_HEIGHT_POS - TILE_SIDE - rowIndex * TILE_SIDE;
 			tiles[columnIndex].emplace_back(xPos, yPos, TILE_SIDE, GetRandomColor());
 		}
 	}
